Move infra stack to us-east-1
Previously it was in Tokyo. Moving to us-east-1 since ACM certs for CloudFront need to be in us-east-1 and I want CDK to manage these certs.

(After destroying the old stacks, need to wait for hours before recreating them since we're essentially move a bucket of a fixed name to a different region, and need the bucket status to propagate globally before the name becomes available again)

__Testing done__

1. `cdk deploy prod`
1. `cdk deploy beta`
1. Update beta name servers
1. `cdk deploy prod`