Add CI/CD workflow (#1)
Previously the deployment was manual. Now we have a workflow that runs
both the deployment to prod and the pre-commit periodically. The
periodic check allows us to catch dependencies that broke beneath our
feet.

__Out of scope__

Testing in main that actually deploys to prod. By default the workflow
deploys to beta.

__Testing done__


https://github.com/boonjiashen/boonjiashen-dot-com-infra/actions/runs/10259041558
Succeeds in beta